requirements/local.txt
asgiref==3.5.0
Django==3.0.5
mysqlclient==2.1.0
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
Unipath==1.1
